% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank_Countries.R
\name{Rank_Countries}
\alias{Rank_Countries}
\title{Rank_countries}
\usage{
Rank_Countries()
}
\value{
plot
}
\description{
Rank_countries() gives back a bar graph showing the countries listed by number of universities in ranking.
}
